bf16 Phase 3b: native CUDA (PTX) - all 6 backends now
PTX has no native bf16 arithmetic (only cvt.*.bf16), so CUDA holds bf16
in an f32 register and converts only at the load/store boundary:

- BasicValueType.BFloat16 maps to a .f32 register (RegisterTypeMapping +
  ParameterTypeRemapping + movement-remap in PTXRegisterAllocator /
  PTXCodeGenerator). RegisterIOTypeRemapping stays .b16 (storage).
- All bf16 arithmetic/compare routes through the f32 tables:
  PTXInstructions.GetArithmeticOperation (unary/binary/ternary) +
  GetCompareOperation remap bf16 -> f32 at the chokepoint.
- ConvertValue bf16<->f32 is a register no-op (Alias); a bf16 endpoint
  against f64/int remaps to f32 then does the real convert.
- Custom Load: ld.global.b16 into a temp .b16 reg, then cvt.f32.bf16.
- Custom Store: cvt.rn.bf16.f32 (round-to-nearest-even) into a temp .b16
  reg, then st.global.b16. EnsureHardwareRegister materializes a bf16
  constant first (cvt cannot take an immediate source).
- bf16 ConstantRegister emits the f32 magnitude immediate.

Round-trip is byte-identical to the emulated backends. cvt.*.bf16 is
native on sm_80+ (Ada/Hopper); verified on sm_89.

PMT PMT_FILTER=BFloat16: all 6 lanes (CPU + CUDA + OpenCL + WebGPU +
WebGPU-NoSubgroups + WebGL + Wasm) 28/0/0. No f16 regression
(PMT_FILTER=Half 181/0/8). Bundle bumps: forks 2.0.18,
SpawnDev.ILGPU 4.13.0-local.2.

+++ ILGPU/Backends/PTX/PTXCodeGenerator.cs
@@ -202,7 +202,7 @@ namespace ILGPU.Backends.PTX
                 string.Empty, "pred",
                 "b8", "b16", "b32", "b64",
                 "f16", "f32", "f64",
-                "f16"); // BFloat16 (ordinal 9) - placeholder; native bf16 PTX codegen is Phase 3
+                "f32"); // BFloat16 (ordinal 9) - bf16 value held in an f32 register (storage uses literal b16)
 
         /// <summary>
         /// Maps basic types to constant-loading target basic types.
@@ -214,7 +214,7 @@ namespace ILGPU.Backends.PTX
                 BasicValueType.Int16, BasicValueType.Int16,
                 BasicValueType.Int32, BasicValueType.Int64,
                 BasicValueType.Int16, BasicValueType.Float32, BasicValueType.Float64,
-                BasicValueType.Int16); // BFloat16 (ordinal 9) - placeholder (Phase 3)
+                BasicValueType.Float32); // BFloat16 (ordinal 9) - f32-register model (mov as f32)
 
         /// <summary>
         /// Maps basic types to constant-loading target basic types.
